+++ packages/content/utils/Icons.ts
@@ -1,96 +1,13 @@
 type NAME_TO_ICON = { [name: string]: string };
+export { getIconForFile, getIconForFolder, getIconForOpenFolder, DEFAULT_FILE, DEFAULT_FOLDER, DEFAULT_FOLDER_OPENED, DEFAULT_ROOT_OPENED, DEFAULT_ROOT } from 'vscode-icons-js';
+import { DEFAULT_FILE } from 'vscode-icons-js';
 
-const folderNamesToIcon = require('../data/FolderNamesToIcon.json') as NAME_TO_ICON;
-const fileExtensions1ToIcon = require('../data/FileExtensions1ToIcon.json') as NAME_TO_ICON;
-const fileExtensions2ToIcon = require('../data/FileExtensions2ToIcon.json') as NAME_TO_ICON;
-const fileNamesToIcon = require('../data/FileNamesToIcon.json') as NAME_TO_ICON;
-const languagesToIcon = require('../data/LanguagesToIcon.json') as NAME_TO_ICON;
 const PBSyntaxesToIcon = require('../data/PastebinSyntaxesToIcon.json') as NAME_TO_ICON;
 
 /**
  * Retrieve url of icon within chrome
  */
 export const getIconUrl = (iconFileName: string) => chrome.runtime.getURL('icons/' + iconFileName);
-
-export const DEFAULT_FOLDER = 'default_folder.svg';
-export const DEFAULT_FOLDER_OPENED = 'default_folder_opened.svg';
-export const DEFAULT_ROOT = 'default_root_folder.svg';
-export const DEFAULT_ROOT_OPENED = 'default_root_folder_opened.svg';
-export const DEFAULT_FILE = 'default_file.svg';
-
-/**
- * Get icon for a folder
- * @param folderName name of folder to find icon for
- * @return icon filename
- */
-export function getIconForFolder(folderName: string) {
-  const folderIcon = folderNamesToIcon[folderName];
-  return folderIcon ? folderIcon : DEFAULT_FOLDER;
-}
-
-let prevExtension: null | string = null;
-let prevIcon: null | string = null;
-
-/**
- * Get icon for a file
- * @param fileName name of file to find icon for
- * @return icon filename
- */
-export function getIconForFile(fileName: string) {
-  // match by exact FileName
-  const iconFromFileName = fileNamesToIcon[fileName];
-  if (iconFromFileName !== undefined) {
-    return iconFromFileName;
-  }
-
-  // match by File Extension
-  const extensions = fileName.split('.');
-  if (extensions.length > 2) {
-    const ext1 = extensions.pop();
-    const ext2 = extensions.pop();
-    // check for `.js.map`, `test.tsx`, ...
-    const iconFromExtension2 = fileExtensions2ToIcon[`${ext2}.${ext1}`];
-    if (iconFromExtension2 !== undefined) {
-      return iconFromExtension2;
-    }
-    // check for `.js`, `tsx`, ...
-
-    if (ext1 === prevExtension) {
-      return prevIcon;
-    }
-    const iconFromExtension1 = fileExtensions1ToIcon[ext1];
-    if (iconFromExtension1 !== undefined) {
-      // memoization
-      prevExtension = ext1;
-      prevIcon = iconFromExtension1;
-      return iconFromExtension1;
-    }
-  } else {
-    const ext = extensions.pop();
-    if (ext === prevExtension) {
-      return prevIcon;
-    }
-    const iconFromExtension = fileExtensions1ToIcon[ext];
-    if (iconFromExtension !== undefined) {
-      // memoization
-      prevExtension = ext;
-      prevIcon = iconFromExtension;
-      return iconFromExtension;
-    }
-  }
-
-  // match by language
-  const fileExtension = fileName.split('.').pop();
-  if (fileExtension !== undefined) {
-    const iconFromLang = languagesToIcon[fileExtension];
-    if (iconFromLang) {
-      return iconFromLang;
-    }
-  }
-
-  // if there's no icon for file, use default one
-  return DEFAULT_FILE;
-}
 
 /**
  * Get icon for a pastebin syntaxes
@@ -104,17 +21,4 @@ export function getIconForPBSyntax(syntaxName: string) {
     return syntaxIcon;
   }
   return DEFAULT_FILE;
-}
-
-/**
- * Get icon for an opened folder
- * @param folderName name of opened folder to icon for
- * @return icon filename
- */
-export function getIconForOpenFolder(folderName: string) {
-  return (
-    getIconForFolder(folderName)
-      .split('.')
-      .shift() + '_opened.svg'
-  );
 }